Fix test failures and deprecation warnings
1. UnifiedWebcamWorkerFlowTest: Use relative timestamps instead of
   hardcoded dates (2026-01-16) in normalizeExifToUtc tests. The
   hardcoded dates caused test failures when the dates became older
   than the 24-hour image retention window.

2. WebcamErrorDetectorTest: Add explicit (int) cast to fix PHP 8.x
   deprecation warning about implicit float-to-int conversion in
   the test image generation callback.

+++ tests/Unit/WebcamErrorDetectorTest.php
@@ -366,7 +366,7 @@ class WebcamErrorDetectorTest extends TestCase
             // Create gradient pattern with good variance
             $r = min(255, ($x * 2) % 256);
             $g = min(255, ($y * 2) % 256);
-            $b = min(255, (($x + $y) * 1.5) % 256);
+            $b = min(255, (int)(($x + $y) * 1.5) % 256);
             
             // Add some structure (edges)
             if (($x % 20) < 10 && ($y % 20) < 10) {
